@@ -3,8 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
+    "axios": "^0.18.0",
+    "react": "^16.8.0",
+    "react-dom": "^16.8.0",
     "react-scripts": "1.0.17"
   },
   "scripts": {
