@@ -1,32 +1,28 @@
-import React, { Component } from 'react';
+import React, { useState, useEffect } from 'react';
+import axios from 'axios';
 import './customers.css';
 
-class Customers extends Component {
-  constructor() {
-    super();
-    this.state = {
-      customers: []
-    };
-  }
+const Customers = () => {
+  const [customers, setCustomers] = useState([]);
 
-  componentDidMount() {
-    fetch('/api/customers')
-      .then(res => res.json())
-      .then(customers => this.setState({customers}, () => console.log('Customers fetched...', customers)));
-  }
 
-  render() {
-    return (
-      <div>
-        <h2>Customers</h2>
-        <ul>
-        {this.state.customers.map(customer => 
+  useEffect(() => {
+
+    axios.get('/api/customers').then((response) => setCustomers(response.data)).catch((error) => console.log(error));
+
+  }, [customers])
+
+
+  return (
+    <div>
+      <h2>Customers</h2>
+      <ul>
+        {customers.map(customer =>
           <li key={customer.id}>{customer.firstName} {customer.lastName}</li>
         )}
-        </ul>
-      </div>
-    );
-  }
-}
+      </ul>
+    </div>
+  );
+};
 
 export default Customers;
